+++ earthkit/__init__.py
@@ -1,4 +1,3 @@
-"""Statistical analysis tools for meteorological and climate data.."""
 
 # Copyright 2022, European Centre for Medium Range Weather Forecasts.
 #
@@ -17,23 +16,8 @@
 try:
     # NOTE: the `version.py` file must not be present in the git repository
     #   as it is generated by setuptools at install time
-    from earthkit.aggregate.version import __version__
+    from earthkit.climate.version import __version__
 except ImportError:  # pragma: no cover
     # Local copy or not installed with setuptools
     __version__ = "999"
 
-
-from earthkit.aggregate import climatology, spatial, temporal
-
-try:
-    from earthkit.data.utils.module_inputs_wrapper import transform_module_inputs
-except ImportError:
-    pass
-else:
-    temporal = transform_module_inputs(temporal)
-
-    climatology = transform_module_inputs(climatology)
-
-    spatial = transform_module_inputs(spatial)
-
-__all__ = ["__version__", "temporal", "climatology", "spatial"]
